refact: melhora na arquitetura do codigo
