+++ src/TreineMais.API/Requests/Auth/AuthRequest.cs
@@ -1,3 +1,3 @@
-namespace TreineMais.Application.DTO.Auth;
+namespace TreineMais.API.Requests.Auth;
 
 public record AuthRequest(string Email, string Password);
