+++ src/TreineMais.API/Requests/Training/TrainingRequest.cs
@@ -1,0 +1,6 @@
+namespace TreineMais.API.Requests.Training;
+
+public record TrainingRequest(
+    string Name,
+    string? Description
+);
