+++ src/TreineMais.API/Requests/User/UserRequest.cs
@@ -1,0 +1,5 @@
+using TreineMais.API.Requests.Profile;
+
+namespace TreineMais.API.Requests.User;
+
+public record UserRequest(string Email, string Password, ProfileRequest Profile);
