--- src/TreineMais.Application/DTO/Auth/LogoutResponse.cs
@@ -1,3 +1,0 @@
-namespace TreineMais.Application.DTO.Auth;
-
-public record LogoutResponse(string Message);
