--- src/TreineMais.Application/DTO/Trainings/TrainingRequest.cs
@@ -1,12 +1,0 @@
-namespace TreineMais.Application.DTO.Trainings;
-
-public record TrainingRequest(
-    string Name,
-    string? Description,
-    TrainingExercisesRequest TrainingExercises
-);
-
-public record TrainingExercisesRequest(
-    int Sets,
-    int Reps,
-    float Weight);
