--- src/TreineMais.Application/DTO/User/UserRequest.cs
@@ -1,5 +1,0 @@
-using TreineMais.Application.DTO.Profile;
-
-namespace TreineMais.Application.DTO.User;
-
-public record UserRequest(string Email, string Password, ProfileRequest Profile);
