--- src/TreineMais.Application/DTO/User/UserResponse.cs
@@ -1,5 +1,0 @@
-using TreineMais.Application.DTO.Profile;
-
-namespace TreineMais.Application.DTO.User;
-
-public record UserResponse(Guid UserId, string Email, ProfileResponse Profile);
