+++ src/TreineMais.Application/Responses/Auth/AuthResponse.cs
@@ -1,3 +1,3 @@
-namespace TreineMais.Application.DTO.Auth;
+namespace TreineMais.Application.Responses.Auth;
 
 public record AuthResponse(string AccessToken, string RefreshToken);
