+++ src/TreineMais.Application/Responses/Auth/LogoutResponse.cs
@@ -1,0 +1,3 @@
+namespace TreineMais.Application.Responses.Auth;
+
+public record LogoutResponse(string Message);
