+++ src/TreineMais.Application/Responses/Training/TrainingResponse.cs
@@ -1,0 +1,6 @@
+namespace TreineMais.Application.Responses.Training;
+
+public record TrainingResponse(
+    int Id,
+    string Name,
+    string? Description);
