+++ src/TreineMais.Application/Responses/User/UserResponse.cs
@@ -1,0 +1,5 @@
+using TreineMais.Application.Responses.Profile;
+
+namespace TreineMais.Application.Responses.User;
+
+public record UserResponse(Guid UserId, string Email, ProfileResponse Profile);
