+++ src/TreineMais.Application/UseCase/ConfirmRefreshToken/RefreshTokenCommand.cs
@@ -1,5 +1,5 @@
 using MediatR;
-using TreineMais.Application.DTO.Auth;
+using TreineMais.Application.Responses.Auth;
 
 namespace TreineMais.Application.UseCase.ConfirmRefreshToken;
